[create sale_type table][add pagination in user ]

+++ database/migrations/2018_11_17_060055_create_tickets_table.php
@@ -15,7 +15,7 @@ class CreateTicketsTable extends Migration
         Schema::create('tickets', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('price');
+            $table->smallInteger('price');
         });
     }
