+++ database/migrations/2018_12_10_085548_create_guide_commissions_table.php
@@ -16,7 +16,7 @@ class CreateGuideCommissionsTable extends Migration
         Schema::create('guide_commissions', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('commission');
-            $table->tinyInteger('ticket_id');
+            $table->smallInteger('ticket_id');
         });
     }
